models, serializers and views are split to separate files: continue; logic from post method of LoginVK class and delete method of LoginVKDeleteToken class is placed into service_objects
