+++ server/gallery_api/gallery_api/rest_api/models/vote.py
@@ -1,8 +1,7 @@
 from django.db import models
 from django.contrib.auth.models import User
-from gallery_api.rest_api.models.photo import Photo
 
 class Vote(models.Model):
     user = models.ForeignKey(User, related_name="votes", on_delete=models.DO_NOTHING)
-    photo = models.ForeignKey(Photo, related_name="votes", on_delete=models.DO_NOTHING)
+    photo = models.ForeignKey("Photo", related_name="votes", on_delete=models.DO_NOTHING)
     vote = models.SmallIntegerField()
