+++ server/gallery_api/gallery_api/rest_api/serializers/login_vk.py
@@ -2,5 +2,6 @@ from rest_framework import serializers
 
 
 class LoginVKSerializer(serializers.Serializer):
-    code = serializers.CharField()
-    network = serializers.CharField()
+    code = serializers.CharField(write_only=True)
+    network = serializers.CharField(write_only=True)
+    token = serializers.CharField(read_only=True)
