+++ server/gallery_api/gallery_api/rest_api/serializers/token.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+
+from rest_framework.authtoken.models import Token
+
+
+class TokenSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Token
+        fields = ["key"]
+
